--- todo.txt
@@ -1,10 +1,0 @@
-mock
-
-Maja:
-
-Iza:
-restart działa tak, że wracamy do wyboru poziomu trudności
-
-
-Pytania:
-Czy możemy losować zdania i usuwać te za długie (zawijanie tekstu jest trudne)
